pyproject.toml
[project]
name = "vg-control"
version = "0.8.0"
description = "Video game control input tracking application"
authors = [{ name = "Wayfarer Labs" }]
requires-python = ">=3.12,<3.13"
readme = "README.md"
dependencies = [
    "python-dotenv>=1.0.0,<2",
    "numpy>=2.0.0,<3",
    "pandas>=2.2.3",
    "psutil>=7.0.0",
    "requests>=2.32.3",
    "tqdm>=4.67.1",
    "GPUtil>=1.4.0",
    "py-cpuinfo>=9.0.0",
    "WMI>=1.5.1 ; sys_platform == 'win32'",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0,<8",
    "black>=23.7.0,<24",
    "flake8>=6.1.0,<7",
    "mypy>=1.5.0,<2",
    "ruff>=0.13.0",
]

[tool.hatch.build.targets.sdist]
include = ["vg_control"]

[tool.hatch.build.targets.wheel]
include = ["vg_control"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 120
target-version = ['py312']

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
